Split tensor extension types into multiple files
